Added jQuery links and basic row loop structure
Note row_chunk.htm.js is a starting point for creating the dynamically
generated row loop. There are a lot of quotes(") that need to be
escaped in javascript strings with \". Before I go and do all that, I
want to 1) decide on what the form elements need in terms of unique
"id", "name", etc and 2) make sure it is more efficient to build the
form via a loop rather than have it static.