Bridge responds! And renamed a bunch of files
…so they fit in the 8.3 arduino file name size limit. Serving these
files from the arduino SD card actually works and the response, though
useless, is real. This took adding a ton of serial debug lines to
figure out how the loop was really working. The main key was the
client.stop() on line 270.

The hard-coded response to the post request are lines 266-269. Yay!