cupcake now saves DIRECT to sd file via PUT request. Export reads both .txt files
This is a good commit. Two ajax request trigger to load program.txt and
settings.txt, which are then put into little text boxes for the user to
copy, paste and save. There should be a link provided for the user to
download the entire thing in one files with a timestamp and description.